remove old logo
