--- src/EdozoLogo2/index.tsx
@@ -1,1 +1,0 @@
-export { EdozoLogo2 } from './EdozoLogo2.tsx';
